requirements.txt
selenium==3.141.0
xlrd==1.2.0
PyYAML==3.13
pymysql
faker
